Refactor logic for creating a payment profile from a Stripe token into a service object
